Power Ups applied to the modified stats

+++ Assets/Project/Scripts/Enemies/IHealth.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IHealth
 {
-    public void GetHit(int amount); 
+    public void GetHit(float amount); 
 }
